CGI::Parse::PSGI::parse_cgi_output - use the HTTP status line code's as the default
i.e. in the case that the "Status:" header is absent, zero, or
otherwise not well-formed.

The default HTTP status code is 200, when the status line is absent,
as returned by HTTP::Response->code.

Note, this alters the behaviour from before in one edge case: when the
Status header is absent, but the HTTP status line is present, and not
200.  Before, the result would be 200 (because only the status header
was checked). Now, the result will be the value of the status
line (because it is checked).  I think this is the most sensible
default, so I am letting it stand.