t/02_parse.t - add test for HTTP status line vs Status header
This just illustrates the current behaviour when both a status line
and a status header is present.

Some non-parsed header CGI scripts emit a status line, and expect that
to be observed.  Note that the status line is not observed, even when
there is no status header (the last test case added).

I suspect this isn't correct behaviour: the status line should be
observed if there is no status header.